+++ praetorian_cli/sdk/chariot.py
@@ -11,6 +11,7 @@ from praetorian_cli.sdk.entities.configurations import Configurations
 from praetorian_cli.sdk.entities.conversations import Conversations
 from praetorian_cli.sdk.entities.credentials import Credentials
 from praetorian_cli.sdk.entities.definitions import Definitions
+from praetorian_cli.sdk.entities.engineer_vm import EngineerVms
 from praetorian_cli.sdk.entities.files import Files
 from praetorian_cli.sdk.entities.hunts import Hunts
 from praetorian_cli.sdk.entities.integrations import Integrations
@@ -66,6 +67,7 @@ class Chariot:
         self.webpage = Webpage(self)
         self.schema = Schema(self)
         self.schedules = Schedules(self)
+        self.vms = EngineerVms(self)
         self.proxy = proxy
 
         if self.proxy == '' and os.environ.get('CHARIOT_PROXY'):